+++ docs/operators-and-hooks-ref.rst
@@ -239,7 +239,7 @@ These integrations allow you to perform various operations within the Microsoft 
    * - `Azure Blob Storage <https://azure.microsoft.com/en-us/services/storage/blobs/>`__
      - :mod:`airflow.providers.microsoft.azure.hooks.azure_storage_blob`
      - :mod:`airflow.providers.microsoft.azure.operators.local_to_azure_storage_blob`,
-       :mod:`airflow.providers.microsoft.azure.operators.azure_storage_blob_download`,
+       :mod:`airflow.providers.microsoft.azure.operators.azure_storage_blob_to_local`,
        :mod:`airflow.providers.microsoft.azure.operators.azure_storage_delete_blob`
      - :mod:`airflow.providers.microsoft.azure.sensors.azure_storage_blob`
